requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dataclasses-json==0.6.6
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.2
langchain-core==0.2.4
langchain-openai==0.1.7
langchain-text-splitters==0.2.1
langsmith==0.1.71
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.31.0
orjson==3.10.3
packaging==23.2
pdfminer.six==20231228
pikepdf==8.15.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyPDF2==3.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sniffio==1.3.1
SQLAlchemy==2.0.30
sympy==1.12.1
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.1
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
